requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
certifi==2025.7.9
    # via requests
charset-normalizer==3.4.2
    # via requests
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.5.1
    # via
    #   huggingface-hub
    #   torch
hf-xet==1.1.5
    # via huggingface-hub
huggingface-hub==0.33.4
    # via
    #   tokenizers
    #   transformers
idna==3.10
    # via requests
jinja2==3.1.6
    # via torch
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
numpy==2.3.1
    # via
    #   torchvision
    #   transformers
nvidia-cublas-cu12==12.6.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.6.80
    # via torch
nvidia-cuda-nvrtc-cu12==12.6.77
    # via torch
nvidia-cuda-runtime-cu12==12.6.77
    # via torch
nvidia-cudnn-cu12==9.5.1.17
    # via torch
nvidia-cufft-cu12==11.3.0.4
    # via torch
nvidia-cufile-cu12==1.11.1.6
    # via torch
nvidia-curand-cu12==10.3.7.77
    # via torch
nvidia-cusolver-cu12==11.7.1.2
    # via torch
nvidia-cusparse-cu12==12.5.4.2
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.3
    # via torch
nvidia-nccl-cu12==2.26.2
    # via torch
nvidia-nvjitlink-cu12==12.6.85
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.6.77
    # via torch
packaging==25.0
    # via
    #   huggingface-hub
    #   transformers
pillow==11.3.0
    # via
    #   -r requirements.in
    #   torchvision
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via transformers
requests==2.32.4
    # via
    #   huggingface-hub
    #   transformers
safetensors==0.5.3
    # via transformers
setuptools==80.9.0
    # via
    #   torch
    #   triton
sympy==1.14.0
    # via torch
tokenizers==0.21.2
    # via transformers
torch==2.7.1
    # via
    #   -r requirements.in
    #   torchvision
torchvision==0.22.1
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.53.2
    # via -r requirements.in
triton==3.3.1
    # via torch
typing-extensions==4.14.1
    # via
    #   huggingface-hub
    #   torch
urllib3==2.5.0
    # via requests
